+++ packages/communication-react/review/beta/communication-react.api.md
@@ -387,7 +387,7 @@ export type CallAdapterClientState = {
     selectedVideoBackgroundEffect?: VideoBackgroundEffect;
     acceptedTransferCallState?: CallState;
     hideAttendeeNames?: boolean;
-    useSounds?: boolean;
+    sounds?: CallingSounds;
 };
 
 // @public
@@ -819,6 +819,11 @@ export type CallingHandlersOptions = {
 
 // @public
 export type CallingReturnProps<Component extends (props: any) => JSX.Element> = GetCallingSelector<Component> extends (state: CallClientState, props: any) => any ? ReturnType<GetCallingSelector<Component>> & Common<CallingHandlers, Parameters<Component>[0]> : never;
+
+// @beta
+export type CallingSounds = {
+    callEnded?: SoundEffect;
+};
 
 // @public
 export interface CallingTheme {
@@ -1741,7 +1746,7 @@ export type CommonCallAdapterOptions = {
     };
     onFetchProfile?: OnFetchProfileCallback;
     soundOptions?: {
-        disableSounds?: boolean;
+        callingSounds?: CallingSounds;
     };
 };
 
@@ -3671,6 +3676,12 @@ export interface SitePermissionsStyles extends BaseCustomStyles {
     primaryButton?: IButtonStyles;
     troubleshootingLink?: ILinkStyles;
 }
+
+// @beta
+export type SoundEffect = {
+    path: string;
+    fileType?: 'mp3' | 'wav' | 'ogg' | 'aac' | 'flac';
+};
 
 // @public
 export interface SpokenLanguageStrings {
