+++ packages/react-composites/review/beta/react-composites.api.md
@@ -242,7 +242,7 @@ export type CallAdapterClientState = {
     selectedVideoBackgroundEffect?: VideoBackgroundEffect;
     acceptedTransferCallState?: CallState;
     hideAttendeeNames?: boolean;
-    useSounds?: boolean;
+    sounds?: CallingSounds;
 };
 
 // @public
@@ -585,6 +585,11 @@ export type CallEndedListener = (event: CallAdapterCallEndedEvent) => void;
 export type CallIdChangedListener = (event: {
     callId: string;
 }) => void;
+
+// @beta
+export type CallingSounds = {
+    callEnded?: SoundEffect;
+};
 
 // @beta
 export type CallParticipantsLocator = {
@@ -1111,7 +1116,7 @@ export type CommonCallAdapterOptions = {
     };
     onFetchProfile?: OnFetchProfileCallback;
     soundOptions?: {
-        disableSounds?: boolean;
+        callingSounds?: CallingSounds;
     };
 };
 
@@ -1742,6 +1747,12 @@ export type Profile = {
 export interface RemoteVideoTileMenuOptions {
     isHidden?: boolean;
 }
+
+// @beta
+export type SoundEffect = {
+    path: string;
+    fileType?: 'mp3' | 'wav' | 'ogg' | 'aac' | 'flac';
+};
 
 // @public
 export type TeamsAdapterOptions = CommonCallAdapterOptions;
