+++ packages/react-composites/src/composites/CallComposite/adapter/index.ts
@@ -70,3 +70,6 @@ export type { CapabilitiesChangedListener } from './CallAdapter';
 
 /* @conditional-compile-remove(teams-identity-support) */
 export type { TeamsCallAdapter } from './CallAdapter';
+
+/* @conditional-compile-remove(calling-sounds) */
+export type { CallingSounds, SoundEffect } from './CallAdapter';
